Add function to expand skip edges (#249)
# Proposed Change

This PR adds a function that checks whether a given skip edge is equivalent to multiple edges connecting the matched source and target vertex. A skip edge `u -> v` is considered equivalent to a path of edges `u' -> ... -> v'` if `(u, u')` and `(v, v')` are matched vertices, and no other nodes on the path have a matching node in the other graph. 

This PR also adds `SKIP_TRUE_POS`, `SKIP_FALSE_NEG` and `SKIP_FALSE_POS` annotations to edges for basic edge errors. The user can relax edge checks by toggling `relax_skips_gt` to check the presence of edges in pred equivalent to a gt skip edge, or `relax_skips_pred` to check the presence of edges in gt equivalent to a pred skip edge, when calling `Metrics.compute`.

This PR does not update the values e.g. precision or recall reported by the `BasicMetrics`, and does not change graph annotations for anything except the `BasicMetrics`.

# Types of Changes
- New feature or enhancement
- Documentation update

Which topics does your change affect? Delete those that do not apply.
- Track Errors

# Checklist
Put an x in the boxes that apply. You can also fill these out after creating the PR. If you're unsure about any of them, don't hesitate to ask. We're here to help! This is simply a reminder of what we are going to look for before merging your code.

- [ ] I have read the developer/contributing docs.
- [ ] I have added tests that prove that my feature works in various situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [ ] I have checked the benchmarking action to verify that my changes did not adversely affect performance.
- [ ] I have written docstrings and checked that they render correctly in the Read The Docs build (created after the PR is opened).
- [ ] I have updated the general documentation including Metric descriptions and example notebooks if necessary.

# Further Comments
If this is a relatively large or complex change, kick off the discussion by explaining why you chose the solution you did and what alternatives you considered, etc...