Incorporate skip edge errors into basic metrics  (#264)
# Proposed Change
- Closes #261
- In order to count errors when skip errors are involved, I added a new
function to look at each edge individually and count it for only one
error class. If we are counting skip errors, the skip flag takes
precedent over any non-skip error flag.
- I added a new test case graph that contains all of the basic errors so
that it was easy to test that the counts were working correctly.
- I changed the way that we compute the total number of edges in
gt/pred. Previously we added tp + fn or tp + fp, but that leads to weird
counts with skip edges. Instead I am now just counting the number of
edges in the graph. I don't think this is exactly what we discussed at
our last meeting, but this feels more correct to me.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- New feature or enhancement

Which topics does your change affect? Delete those that do not apply.
- Metrics

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the developer/contributing docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [x] I have checked that I maintained or improved code coverage.
- [x] I have checked the benchmarking action to verify that my changes
did not adversely affect performance.
- [x] I have written docstrings and checked that they render correctly
in the Read The Docs build (created after the PR is opened).
- [x] I have updated the general documentation including Metric
descriptions and example notebooks if necessary.