Add TF measure to track_overlap (#259)
# Proposed Change
Closes #40

This PR adds an additional result to the `TrackOverlap` metrics
representing track fractions (TF): the **unweighted** track
effectiveness.

This is calculated as the average fraction of maximally overlapping
predicted tracklets onto the GT tracklets. Unlike track effectiveness,
this metric does not take into account the tracklet length when
computing fractions.

Unlike track purity/track effectiveness, it's not trivial to "invert"
the track fractions number, so I only test the "forward" direction in
this case. I also got a couple of floating point errors in the tests, so
I changed the assertion to `np.allclose`.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- New feature or enhancement

Which topics does your change affect? Delete those that do not apply.
- Metrics

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the developer/contributing docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [x] I have checked that I maintained or improved code coverage.
- [x] I have checked the benchmarking action to verify that my changes
did not adversely affect performance.
- [x] I have written docstrings and checked that they render correctly
in the Read The Docs build (created after the PR is opened).
- [x] I have updated the general documentation including Metric
descriptions and example notebooks if necessary.

---------

Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>