Implement cell cycle accuracy metric (#247)
# Proposed Matcher or Metric Addition
- Metric

Closes #41 

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the developer/contributing docs.
- [x] I have added tests for the standard test examples documented
[here](https://traccuracy.readthedocs.io/en/latest/test_cases/index.html),
along with end-to-end tests.
- [ ] I have checked that I maintained or improved code coverage.
- [x] I have added benchmarking functions for my change
`tests/bench.py`.
- [x] I have added a page to the documentation with a complete
description of my matcher/metric including any references.
- [x] I have written docstrings and checked that they render correctly
in the Read The Docs build (created after the PR is opened).

---------

Co-authored-by: Draga Doncila Pop <17995243+DragaDoncila@users.noreply.github.com>
Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>