Allow skip edges in divisions (#253)
# Proposed Change
Closes #214 and #262 as part of the effort in #249 to expand support for
skip edges in our metrics.

Changes to `_classify_divisions`:
- In order to handle both skip and normal edges, we now check if out
going edges from a division node are tps or skip_tps instead of checking
if the daughters match.
- This change adds the requirement that graphs be annotated with basic
errors prior to division annotation, but we handle that automatically if
it's not already done.

Changes to `_correct_shifted_divisions`:
- Previously we would look at predecessors/successors in a particular
frame and check that the nodes had a match in the other graph. This
becomes very complicated with skip edges.
- The new strategy is to take the division node in the earliest frame
and walk the graph until we find daughters that are true positive nodes.
For each valid daughter, we check that on the opposite graph there is a
path between the match to this daughter and the match to the division
node.
- This works agnostically for both normal and skip edge cases.

Updates division metrics to handle counting tp skip divisions

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- New feature or enhancement

Which topics does your change affect? Delete those that do not apply.
- Track Errors
- Metrics

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the developer/contributing docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [x] I have checked that I maintained or improved code coverage.
- [x] I have checked the benchmarking action to verify that my changes
did not adversely affect performance.
- [x] I have written docstrings and checked that they render correctly
in the Read The Docs build (created after the PR is opened).
- [x] I have updated the general documentation including Metric
descriptions and example notebooks if necessary.

# Further Comments
If this is a relatively large or complex change, kick off the discussion
by explaining why you chose the solution you did and what alternatives
you considered, etc...

---------

Co-authored-by: Caroline Malin-Mayor <cmalinmayor@gmail.com>
Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>
Co-authored-by: Draga Doncila Pop <17995243+DragaDoncila@users.noreply.github.com>