Documentation updates (#265)
# Proposed Change
I started this PR just to deal with skip edge related docs changes but
it ended up involving a long of miscellaneous things
- Updates to the docs regarding skip edges and divisions are in open PR
#253
- Move glossary into a separate documentation page, add terms to close
#76 and cross-reference terms when useful
- Fully open to requests for other terms to be added. I went off the
list we had in the open issue and then also added some others that I
thought were useful. But I'm sure there is more that could go in.
- Cross reference between metrics and the errors that they use
- Close #260 
    - Remove references to "gap closing"
    - Add annotated examples of skip edges to basic errors docs
- Remove stand alone gap closing/skip edge docs page, instead define in
glossary and reference as needed on specific error/metrics pages
- Closes #163 
- Closes #75 
- Adds a featured works section with links to papers that use traccuracy
- Closes #266 by toggling "inherited-members" in autoapi config
- Move Results and Matched into their own files to resolve a docs cross
referencing issue
- Closes #269 

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- Documentation update

Which topics does your change affect? Delete those that do not apply.
- Track Errors
- Metrics

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [ ] I have read the developer/contributing docs.
- [ ] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [ ] I have checked the benchmarking action to verify that my changes
did not adversely affect performance.
- [ ] I have written docstrings and checked that they render correctly
in the Read The Docs build (created after the PR is opened).
- [ ] I have updated the general documentation including Metric
descriptions and example notebooks if necessary.

# Further Comments
If this is a relatively large or complex change, kick off the discussion
by explaining why you chose the solution you did and what alternatives
you considered, etc...

---------

Co-authored-by: Draga Doncila Pop <17995243+DragaDoncila@users.noreply.github.com>