Ensure point loader always returns positive integer IDs (#283)
# Proposed Change

This PR updates the points loader to ensure that the returned graph
contains positive integer node IDs.

We did look at potentially regenerating the node IDs if they do not
match the requirements, but it would be a huge headache to do this and
still keep correspondence with the parent IDs so we decided to simply
error.

# Types of Changes
- Maintenance (e.g. dependencies, CI, releases, etc.)

Which topics does your change affect? Delete those that do not apply.
- Loaders

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [ ] I have read the developer/contributing docs.
- [ ] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [ ] I have checked the benchmarking action to verify that my changes
did not adversely affect performance.
- [ ] I have written docstrings and checked that they render correctly
in the Read The Docs build (created after the PR is opened).
- [ ] I have updated the general documentation including Metric
descriptions and example notebooks if necessary.

Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>
Co-authored-by: Morgan Schwartz <msschwartz21@gmail.com>