adding CHOTA metric (#274)
If you are implementing a new matcher or metric, please append this
`&template=new_matcher_metric.md` to your url to load the correct
template.

# Proposed Change
Adding CHOTA metric, reference https://arxiv.org/pdf/2408.11571

# Types of Changes
- New feature or enhancement

Which topics does your change affect? Delete those that do not apply.
- Metrics

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [ ] I have read the developer/contributing docs.
- [ ] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [ ] I have checked the benchmarking action to verify that my changes
did not adversely affect performance.
- [ ] I have written docstrings and checked that they render correctly
in the Read The Docs build (created after the PR is opened).
- [ ] I have updated the general documentation including Metric
descriptions and example notebooks if necessary.

# Further Comments
Any suggestion on how to test this?

---------

Co-authored-by: Morgan Schwartz <msschwartz21@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>