Raise error if only one graph has error annotations (#294)
# Proposed Change

This PR checks that error classification functions are being called with
either both graphs having error annotations, or neither graph having
error annotations. This ensures a user cannot end up with inconsistent
annotations on graphs by accidentally re-running classification on a
graph that already has errors annotated e.g. by keeping GT the same and
only changing pred.

In future we may want to add a utility that "cleans" an annotated graph
so that a user can reuse the graph without requiring a copy.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- Bugfix (non-breaking change which fixes an issue)

Which topics does your change affect? Delete those that do not apply.
- Track Errors

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the developer/contributing docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [ ] I have checked the benchmarking action to verify that my changes
did not adversely affect performance.
- [ ] I have written docstrings and checked that they render correctly
in the Read The Docs build (created after the PR is opened).
- [ ] I have updated the general documentation including Metric
descriptions and example notebooks if necessary.

---------

Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>