Add tests and docs for TF measure (#297)
# Proposed Change

Closes #225, #164

This PR adds tests and docs for the TF measure (unweighted track
effectiveness). Currently there are only 6 test cases where the TF value
differs to the TE value, as we can only expect differences there when
there is more than one GT track and the errors aren't symmetrical across
tracks. @msschwartz21 @cmalinmayor let me know if you think this is
sufficient or if we want to add a couple more tests for which TF is
different to TE.

# Types of Changes

- Documentation update
- Tests and benchmarks

- Metrics

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the developer/contributing docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [x] I have checked that I maintained or improved code coverage.
- [x] I have checked the benchmarking action to verify that my changes
did not adversely affect performance.
- [x] I have written docstrings and checked that they render correctly
in the Read The Docs build (created after the PR is opened).
- [x] I have updated the general documentation including Metric
descriptions and example notebooks if necessary.

---------

Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>
Co-authored-by: Morgan Schwartz <msschwartz21@gmail.com>