docs: :memo: Add metrics overview table and link to in depth descriptions in overviews (#302)
# Proposed Change
Closes #299, in preparation for students at AI-MBL to explore the
available matchers and metrics.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- Documentation update

Which topics does your change affect? Delete those that do not apply.
- Matchers
- Metrics

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the developer/contributing docs.
- [ ] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [ ] I have checked the benchmarking action to verify that my changes
did not adversely affect performance.
- [x] I have written docstrings and checked that they render correctly
in the Read The Docs build (created after the PR is opened).
- [x] I have updated the general documentation including Metric
descriptions and example notebooks if necessary.