Improve error test case coverage (#303)
Closes #254 

A few things that I'm deciding we don't need to test, but let me know if
you disagree
- CTC (anything with shifted divisions)
- Basic (division cases b/c now implicitly tested in division errors
since we use node/edge errors; shifted division cases)
- Divisions (basic node/edge errors, gap close w/o divisions, not
one-to-one)

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- Tests and benchmarks

Which topics does your change affect? Delete those that do not apply.
- Track Errors

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the developer/contributing docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [x] I have checked that I maintained or improved code coverage.
- [x] I have checked the benchmarking action to verify that my changes
did not adversely affect performance.
- [x] I have written docstrings and checked that they render correctly
in the Read The Docs build (created after the PR is opened).
- [x] I have updated the general documentation including Metric
descriptions and example notebooks if necessary.

# Further Comments
If this is a relatively large or complex change, kick off the discussion
by explaining why you chose the solution you did and what alternatives
you considered, etc...