feat: :sparkles: Complete tracks metric (#312)
# Proposed Metric Addition
Closes #39 by adding the complete tracks metric

Questions for reviewers:
- I tested our one larger example. Other end to end tests needed?
- I didn't allow a larger frame buffer than 0. Should we? I looked into
a bit and it requires a good amount of discussion/reimplementation,
since an offset division also makes a bunch of false positive and false
negative nodes/edges that we would need to ignore, at least for the
"lineage correct" metric. For tracklets it does't really make sense to
correct shifted divisions, although it would then make a weird
inconsistency between the definitions if we did correct for the
lineages.
- What happened to our CodeCov bot? (I did run it locally and get to
100% coverage of my added metric file)

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the developer/contributing docs.
- [x] I have added tests for the standard test examples documented
[here](https://traccuracy.readthedocs.io/en/latest/test_cases/index.html),
along with end-to-end tests.
- [x] I have checked that I maintained or improved code coverage.
- [x] I have added benchmarking functions for my change
`tests/bench.py`.
- [x] I have added a page to the documentation with a complete
description of my matcher/metric including any references.
- [x] I have written docstrings and checked that they render correctly
in the Read The Docs build (created after the PR is opened).

# Further Comments
Things that tripped me up worth discussing:
- CTC error computation doesn't annotate TP edges
- What exactly the pyctc metrics does is not entirely clear to me.
https://github.com/CellTrackingChallenge/py-ctcmetrics/blob/3c87a5b9a1ead4018601ba5721bdb8c0238889b2/ctc_metrics/utils/representations.py#L380
- I think this means if they have a 2-to-1 matching (e.g. a non split
node), they remove the second one they come across, to make the matching
1-to-1? In my implementation, I count all Non-Split Nodes as errors,
which is potentially slightly harsher, but seems more correct.

---------

Co-authored-by: Morgan Schwartz <msschwartz21@gmail.com>