Add tracking graph method for clearing annotations on the graph (#306)
Closes #295

My approach to this is to create a new method on the tracking graph that
first loops through all nodes and edges and pops off any NodeFlag or
EdgeFlag items in the attributes. ~~Next I wanted to reset all of the
attributes on TrackingGraph that we set during initialization which I'm
currently doing by rerunning `__init__`. That feels a bit sketch and
ruff isn't super happy about it. @DragaDoncila @cmalinmayor What do you
guys think? Should we just pull out what's in the init into a separate
function that makes it easier to rerun and reset those values?~~ Updated
to use a helper function that is called in the init to set all of those
values and can be rerun after clearing the graph.

Note: I fixed the typing issues in #305 so I'll wait until that PR is
merged first

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- New feature or enhancement

Which topics does your change affect? Delete those that do not apply.
- Core functionality (e.g. `TrackingGraph`, `run_metrics`, `cli`, etc.)

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the developer/contributing docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [x] I have checked that I maintained or improved code coverage.
- [x] I have checked the benchmarking action to verify that my changes
did not adversely affect performance.
- [x] I have written docstrings and checked that they render correctly
in the Read The Docs build (created after the PR is opened).
- [x] I have updated the general documentation including Metric
descriptions and example notebooks if necessary.

# Further Comments
If this is a relatively large or complex change, kick off the discussion
by explaining why you chose the solution you did and what alternatives
you considered, etc...