Use pylapy for solving the LAP (sparse) in PointMatcher/IouMatcher (#320)
It resolves #319

PointMatcher was not solving the correct extended LAP problem. Now it
does, and node labeling do not impact the metric any longer (except if
two costs are strictly equals).

In addition, the LAP is solved using scipy csgraph instead of
linear_sum_assignment, as the LAP is sparse, it is much faster.

I also modified IOUMatcher for its one-to-one assignments, so that it
similarly relies on the sparse solving from pylapy.

Which topics does your change affect? Delete those that do not apply.
- Matchers (PointMatcher, IOUMatcher)

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the developer/contributing docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [x] I have checked that I maintained or improved code coverage.
- [x] I have checked the benchmarking action to verify that my changes
did not adversely affect performance.
- [x] I have written docstrings and checked that they render correctly
in the Read The Docs build (created after the PR is opened).
- [x] I have updated the general documentation including Metric
descriptions and example notebooks if necessary.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>