Pass label key to overlap compute (#325)
# Proposed Change
Noticed that when computing the bbox and labels, we were strictly
looking for `"segmentation_id"` rather than passing the graph's
`label_key` in. This PR updates the `graph_bbox_and_labels` function to
use the `label_key`, and adds tests that were failing with main, but
pass with this change.

# Types of Changes
- Bugfix (non-breaking change which fixes an issue)

Which topics does your change affect? Delete those that do not apply.
- Matchers

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [ ] I have read the developer/contributing docs.
- [ ] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [ ] I have checked the benchmarking action to verify that my changes
did not adversely affect performance.
- [ ] I have written docstrings and checked that they render correctly
in the Read The Docs build (created after the PR is opened).
- [ ] I have updated the general documentation including Metric
descriptions and example notebooks if necessary.

---------

Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>