Fix argument to `_get_mbc` (#326)
# Proposed Change
Closes #321.

We were accidentally passing the wrong value through to the `get_mbc`
function during the frame buffer checks, giving us the strange 0.5
reported in #321. Added a test that passes with PR and fails on main.

# Types of Changes
- Bugfix (non-breaking change which fixes an issue)

- Metrics

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [ ] I have read the developer/contributing docs.
- [ ] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [ ] I have checked the benchmarking action to verify that my changes
did not adversely affect performance.
- [ ] I have written docstrings and checked that they render correctly
in the Read The Docs build (created after the PR is opened).
- [ ] I have updated the general documentation including Metric
descriptions and example notebooks if necessary.

---------

Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>