Split error annotations into basic and ctc (#324)
# Proposed Change
Partially addresses #321 

Prior to this PR, we were marking `node` and `edge` errors as annotated
when any metric was computed, so running basic then CTC or vice-versa
led to one of the metrics being incorrect, because `basic` errors and
`CTC` errors are not the same flags.

This PR splits the flags into `basic_node/edge_errors` (used by all
metrics except CTC) and `ctc_node/edge_errors` (only used by CTC).

# Types of Changes
- Bugfix (non-breaking change which fixes an issue)

Which topics does your change affect? Delete those that do not apply.
- Track Errors
- Metrics

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [ ] I have read the developer/contributing docs.
- [ ] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [ ] I have checked the benchmarking action to verify that my changes
did not adversely affect performance.
- [ ] I have written docstrings and checked that they render correctly
in the Read The Docs build (created after the PR is opened).
- [ ] I have updated the general documentation including Metric
descriptions and example notebooks if necessary.

---------

Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>
Co-authored-by: Caroline Malin-Mayor <cmalinmayor@gmail.com>